hikey960: create dedicated .rc files for power parameters
Hikey960 has different CPU architecture with Hikey, these two platforms
should use their own dedicate configurations for power, especially for
EAS related settings.

This patch is to split the power related settings into two files, one
is init.hikey.power.rc and another is init.hikey960.power.rc. The main
different setting for power setting is related with EAS parameters in
proc nodes.

For Hikey we set the parameters as below:

    write /proc/sys/kernel/sched_is_big_little 0
    write /proc/sys/kernel/sched_cstate_aware 1
    write /proc/sys/kernel/sched_sync_hint_enable 1
    write /proc/sys/kernel/sched_initial_task_util 0

For Hikey960 we set the parameters as below:

    write /sys/kernel/debug/sched_features "ENERGY_AWARE"
    write /proc/sys/kernel/sched_is_big_little 1
    write /proc/sys/kernel/sched_cstate_aware 1
    write /proc/sys/kernel/sched_sync_hint_enable 0
    write /proc/sys/kernel/sched_initial_task_util 0

On Hikey960 we also set default boost margin to 50%, and
touch-boost to 100%. This will let the CPU frequency to reach
highest OPP so can reduce scheduling latency; this is used to
improve the GPU rendering performance.

    write /dev/stune/top-app/schedtune.boost 50
    write /dev/stune/foreground/schedtune.boost 50

Change-Id: I641f5c78075012153fb73e830cb31dfc8dfd6e87
Signed-off-by: Leo Yan <leo.yan@linaro.org>
[jstultz: Added and tweaked some of the parameters to work with
 the parameterized powerHAL]
Signed-off-by: John Stultz <john.stultz@linaro.org>