hikey: Add HIKEY_USE_LEGACY_TI_BLUETOOTH build option to transition to common bluetooth HAL
This patch, a reworked version of changes from Satish Patel
adds a build option to enable the common generic linux bluetooth
HAL vs the legacy TI HAL.

The legacy TI HAL is only needed on 4.4 and older kernels.

NOTE: One needs to be using a kernel that supports the standard
linux hci0 interface to successfully use this.

Change-Id: I74927aa651b184fdb65957fcb54c4c0afffbfa7c
Signed-off-by: John Stultz <john.stultz@linaro.org>