hikey960: Migrate hikey960 build to use Image.gz-dtb kernels
For the UEFI transition we need to move to Image.gz-dtb kernels,
instead of using the separate dts partition as used w/ the HiSi
bootloader.

This patch depends on a Image.gz-dtb-hikey960-4.9/4 kernel to be
in hikey-kernel, but can be used with HiSi bootloader as long
as the dt.img is still flashed.

Change-Id: I8c9ab8a7df516c84c6c5dac57e80388d4d2e40a0
Signed-off-by: John Stultz <john.stultz@linaro.org>