powerHAL: Add interface for GPU and DDR performance boosting
For UI performance we need to ensure GPU and DDR frequency to meet the
performance requirement, otherwise even set the CPU boost margin to high
value but the bottleneck is GPU or DDR side but not CPU. So this commit
introduces the interface to support GPU and DDR performance boosting.

After profiling on Hikey960, this patch adds the related performance
boosting parameters for it. The DDR is set to 685MHz as minimum
frequency and GPU is set to 960MHz as minimum value.

Change-Id: Id5f6b0605a90f76cf55d891f4b7df3c0bf193aee
Signed-off-by: Leo Yan <leo.yan@linaro.org>