powerHAL: Hikey960: Reduce schedtune boost margin to 20%
According to the profiling on Hikey960, usually we need set 15% ~ 20%
boost margin for schedtune for UI performance boosting. With this
setting and using PELT signal, the janks% can be decreased to <30%.

This patch is to set boost margin to 20% and this also can leave the
room for PELT performance optimization.

Change-Id: Ic5eaa78294248076a0e2310461cd00a2fd7884e1
Signed-off-by: Leo Yan <leo.yan@linaro.org>